+++ electron/app/index.ts
@@ -7,7 +7,6 @@ import Nucleus from 'nucleus-nodejs';
 import yargs from 'yargs';
 import {hideBin} from 'yargs/helpers';
 
-import {fixPath} from '@utils/path';
 import '@utils/segment';
 
 import terminal from '@root/cli/terminal';
@@ -16,6 +15,7 @@ import './git/ipc';
 import './ipc/ipcListeners';
 import {openApplication} from './openApplication';
 import {initNucleus, saveInitialK8sSchema, setDeviceID, setProjectsRootFolder} from './utils';
+import {fixPath} from './utils/path';
 
 const isDev = process.env.NODE_ENV === 'development';
