+++ shared/src/utils/index.ts
@@ -1,3 +1,5 @@
 export * from './electronStore';
+export * from './env';
 export * from './fileSystem';
+export * from './hotkey';
 export * from './selectors';
