+++ src/navsections/KustomizationSectionBlueprint/KustomizationQuickAction.tsx
@@ -13,12 +13,12 @@ import {restartPreview, startPreview, stopPreview} from '@redux/services/preview
 
 import {QuickActionCompare, QuickActionPreview} from '@components/molecules';
 
-import {defineHotkey} from '@utils/defineHotkey';
 import {isDefined} from '@utils/filter';
 import {isResourcePassingFilter} from '@utils/resources';
 
 import {hotkeys} from '@monokle-desktop/shared/constants/hotkeys';
 import {ItemCustomComponentProps} from '@monokle-desktop/shared/models';
+import {defineHotkey} from '@monokle-desktop/shared/utils/hotkey';
 
 import * as S from './KustomizationQuickAction.styled';
